requirements.txt
altgraph==0.17.4
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
Cython==3.0.9
google-auth==2.28.2
google-auth-oauthlib==1.2.0
idna==3.6
keyboard==0.13.5
macholib==1.16.3
MouseInfo==0.1.3
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.10
PyMsgBox==1.0.9
pynput==1.7.6
pyobjc==10.1
pyobjc-core==10.1
pyobjc-framework-Accessibility==10.1
pyobjc-framework-Accounts==10.1
pyobjc-framework-AddressBook==10.1
pyobjc-framework-AdServices==10.1
pyobjc-framework-AdSupport==10.1
pyobjc-framework-AppleScriptKit==10.1
pyobjc-framework-AppleScriptObjC==10.1
pyobjc-framework-ApplicationServices==10.1
pyobjc-framework-AppTrackingTransparency==10.1
pyobjc-framework-AudioVideoBridging==10.1
pyobjc-framework-AuthenticationServices==10.1
pyobjc-framework-AutomaticAssessmentConfiguration==10.1
pyobjc-framework-Automator==10.1
pyobjc-framework-AVFoundation==10.1
pyobjc-framework-AVKit==10.1
pyobjc-framework-AVRouting==10.1
pyobjc-framework-BackgroundAssets==10.1
pyobjc-framework-BusinessChat==10.1
pyobjc-framework-CalendarStore==10.1
pyobjc-framework-CallKit==10.1
pyobjc-framework-CFNetwork==10.1
pyobjc-framework-Cinematic==10.1
pyobjc-framework-ClassKit==10.1
pyobjc-framework-CloudKit==10.1
pyobjc-framework-Cocoa==10.1
pyobjc-framework-Collaboration==10.1
pyobjc-framework-ColorSync==10.1
pyobjc-framework-Contacts==10.1
pyobjc-framework-ContactsUI==10.1
pyobjc-framework-CoreAudio==10.1
pyobjc-framework-CoreAudioKit==10.1
pyobjc-framework-CoreBluetooth==10.1
pyobjc-framework-CoreData==10.1
pyobjc-framework-CoreHaptics==10.1
pyobjc-framework-CoreLocation==10.1
pyobjc-framework-CoreMedia==10.1
pyobjc-framework-CoreMediaIO==10.1
pyobjc-framework-CoreMIDI==10.1
pyobjc-framework-CoreML==10.1
pyobjc-framework-CoreMotion==10.1
pyobjc-framework-CoreServices==10.1
pyobjc-framework-CoreSpotlight==10.1
pyobjc-framework-CoreText==10.1
pyobjc-framework-CoreWLAN==10.1
pyobjc-framework-CryptoTokenKit==10.1
pyobjc-framework-DataDetection==10.1
pyobjc-framework-DeviceCheck==10.1
pyobjc-framework-DictionaryServices==10.1
pyobjc-framework-DiscRecording==10.1
pyobjc-framework-DiscRecordingUI==10.1
pyobjc-framework-DiskArbitration==10.1
pyobjc-framework-DVDPlayback==10.1
pyobjc-framework-EventKit==10.1
pyobjc-framework-ExceptionHandling==10.1
pyobjc-framework-ExecutionPolicy==10.1
pyobjc-framework-ExtensionKit==10.1
pyobjc-framework-ExternalAccessory==10.1
pyobjc-framework-FileProvider==10.1
pyobjc-framework-FileProviderUI==10.1
pyobjc-framework-FinderSync==10.1
pyobjc-framework-FSEvents==10.1
pyobjc-framework-GameCenter==10.1
pyobjc-framework-GameController==10.1
pyobjc-framework-GameKit==10.1
pyobjc-framework-GameplayKit==10.1
pyobjc-framework-HealthKit==10.1
pyobjc-framework-ImageCaptureCore==10.1
pyobjc-framework-InputMethodKit==10.1
pyobjc-framework-InstallerPlugins==10.1
pyobjc-framework-InstantMessage==10.1
pyobjc-framework-Intents==10.1
pyobjc-framework-IntentsUI==10.1
pyobjc-framework-IOBluetooth==10.1
pyobjc-framework-IOBluetoothUI==10.1
pyobjc-framework-IOSurface==10.1
pyobjc-framework-iTunesLibrary==10.1
pyobjc-framework-KernelManagement==10.1
pyobjc-framework-LatentSemanticMapping==10.1
pyobjc-framework-LaunchServices==10.1
pyobjc-framework-libdispatch==10.1
pyobjc-framework-libxpc==10.1
pyobjc-framework-LinkPresentation==10.1
pyobjc-framework-LocalAuthentication==10.1
pyobjc-framework-LocalAuthenticationEmbeddedUI==10.1
pyobjc-framework-MailKit==10.1
pyobjc-framework-MapKit==10.1
pyobjc-framework-MediaAccessibility==10.1
pyobjc-framework-MediaLibrary==10.1
pyobjc-framework-MediaPlayer==10.1
pyobjc-framework-MediaToolbox==10.1
pyobjc-framework-Metal==10.1
pyobjc-framework-MetalFX==10.1
pyobjc-framework-MetalKit==10.1
pyobjc-framework-MetalPerformanceShaders==10.1
pyobjc-framework-MetalPerformanceShadersGraph==10.1
pyobjc-framework-MetricKit==10.1
pyobjc-framework-MLCompute==10.1
pyobjc-framework-ModelIO==10.1
pyobjc-framework-MultipeerConnectivity==10.1
pyobjc-framework-NaturalLanguage==10.1
pyobjc-framework-NetFS==10.1
pyobjc-framework-Network==10.1
pyobjc-framework-NetworkExtension==10.1
pyobjc-framework-NotificationCenter==10.1
pyobjc-framework-OpenDirectory==10.1
pyobjc-framework-OSAKit==10.1
pyobjc-framework-OSLog==10.1
pyobjc-framework-PassKit==10.1
pyobjc-framework-PencilKit==10.1
pyobjc-framework-PHASE==10.1
pyobjc-framework-Photos==10.1
pyobjc-framework-PhotosUI==10.1
pyobjc-framework-PreferencePanes==10.1
pyobjc-framework-PushKit==10.1
pyobjc-framework-Quartz==10.1
pyobjc-framework-QuickLookThumbnailing==10.1
pyobjc-framework-ReplayKit==10.1
pyobjc-framework-SafariServices==10.1
pyobjc-framework-SafetyKit==10.1
pyobjc-framework-SceneKit==10.1
pyobjc-framework-ScreenCaptureKit==10.1
pyobjc-framework-ScreenSaver==10.1
pyobjc-framework-ScreenTime==10.1
pyobjc-framework-ScriptingBridge==10.1
pyobjc-framework-SearchKit==10.1
pyobjc-framework-Security==10.1
pyobjc-framework-SecurityFoundation==10.1
pyobjc-framework-SecurityInterface==10.1
pyobjc-framework-SensitiveContentAnalysis==10.1
pyobjc-framework-ServiceManagement==10.1
pyobjc-framework-SharedWithYou==10.1
pyobjc-framework-SharedWithYouCore==10.1
pyobjc-framework-ShazamKit==10.1
pyobjc-framework-Social==10.1
pyobjc-framework-SoundAnalysis==10.1
pyobjc-framework-Speech==10.1
pyobjc-framework-SpriteKit==10.1
pyobjc-framework-StoreKit==10.1
pyobjc-framework-Symbols==10.1
pyobjc-framework-SyncServices==10.1
pyobjc-framework-SystemConfiguration==10.1
pyobjc-framework-SystemExtensions==10.1
pyobjc-framework-ThreadNetwork==10.1
pyobjc-framework-UniformTypeIdentifiers==10.1
pyobjc-framework-UserNotifications==10.1
pyobjc-framework-UserNotificationsUI==10.1
pyobjc-framework-VideoSubscriberAccount==10.1
pyobjc-framework-VideoToolbox==10.1
pyobjc-framework-Virtualization==10.1
pyobjc-framework-Vision==10.1
pyobjc-framework-WebKit==10.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySide6==6.6.1
PySide6-Addons==6.6.1
PySide6-Essentials==6.6.1
pytweening==1.2.0
requests==2.31.0
requests-oauthlib==1.4.0
rsa==4.9
rubicon-objc==0.4.7
screeninfo==0.8.1
setuptools==69.0.2
shiboken6==6.6.1
six==1.16.0
SQLAlchemy==2.0.28
typing_extensions==4.10.0
urllib3==2.2.1
